tts-engine/requirements.txt
chatterbox-tts
torch>=2.2.0
torchaudio>=2.2.0
